+++ CHANGELOG.md
@@ -30,6 +30,7 @@
 * Improve "Skip unmapped pages" by detecting uncommitted segments
 * Cleaner state saving code and compress with stream instead of packets
 * Queue saving memory pages for less write() calls
+* Queue saving compressed memory pages to be saved in temporary memory segment
 
 ### Fixed